feat: Add team listing screen and navigation
Create a new Jetpack Compose screen, `ListadoScreen`, to display a list of teams. This screen includes a styled title "Equipos".

A navigation graph, `AppNavGraph`, is introduced to manage navigation, setting up the route for the new `ListadoScreen`. The `MainActivity` is updated to use this navigation graph.

This commit also includes:
*   A new data class `Equipo` to model team data with Firebase Firestore annotations.
*   The addition of Firebase Firestore, Navigation Compose, and Coil dependencies.
*   An update to `RegistrerLigaScreen` to accept a `NavHostController`.
*   A new color definition, `BlueJJ`.